requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==23.12.1
click==8.1.7
colorama==0.4.6
decouple==0.0.7
Django==4.2.6
django-debug-toolbar==4.2.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
